net/designware: fix phy_addr type to int
so we can pass -1 for auto detect

+++ include/net/designware.h
@@ -2,7 +2,7 @@
 #define __DWC_UNIMAC_H
 
 struct dwc_ether_platform_data {
-	u8 phy_addr;
+	int phy_addr;
 	void (*fix_mac_speed)(int speed);
 };
